Implemented accurate bullet/asteroids collision
I'm exploiting the fact that each asteroid is made up of a bunch of
points around a center point, which means you can think of every edge of
the asteroid as being a side of a triangle that consists of the two
points on that edge and the center point. This makes collision detection
a matter of detecting whether or not a point (the bullet's position) is
inside any of these triangles.